\name{FastSparseGRM-package}
\alias{FastSparseGRM-package}
\alias{FastSparseGRM}
\docType{package}
\title{
\packageTitle{FastSparseGRM}
}
\description{
\packageDescription{FastSparseGRM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{FastSparseGRM}
\packageIndices{FastSparseGRM}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{FastSparseGRM}

Maintainer: \packageMaintainer{FastSparseGRM}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
